+++ data/duke.txt
@@ -1,3 +1,4 @@
 T | 1 | borrow book
 D | 1 | return book | 2nd of December 2019, 6pm
 E | 0 | project meeting | 6th of June 2019, 12am
+T | 0 | shawn
