+++ src/main/java/Command.java
@@ -1,0 +1,6 @@
+public abstract class Command {
+
+    public abstract void execute(TaskList tasks, Storage storage);
+
+    public abstract boolean isExit();
+}
